Split out {Odometry,Vision}{Supplier,Receiver}
The goal of this architecture is that we can have multiple instances
of GreyPoseEstimator that each consume different providers. We want
to be able to compare them in AdvantageScope and decide what combination
of suppliers we want for pose estimation. This is the first pass at
that --- there is a GreyPoseEstimator that just does what the old one
used to do with wheel odometry alone. There is also a new fused instance
which attempts to pull data from the rear limelight.