Fix odometry crash and log (not crash) when exceptions occur
Currently when an exception occurs in the odometry thread, we just stop
the thread and odometry stops updating forever.

With this change, we actually log the exception and let the thread
continue. This depends on where exactly the exception occurs but
most interesting exceptions will occur when propagating
odometry updates from the GreyPoseEstimator type and into the rest of
the robot.

Brought in a subset of the CrashTracker from last year.

Also fixed a crash that we observed last weekend. This was the
exception that we saw:

```
OdometrySupplier thread died.
Cannot invoke "choreo.trajectory.Trajectory.sampleAt(double, boolean)" because "this.m_trajectory" is null
[Ljava.lang.StackTraceElement;@1caa30d
java.lang.NullPointerException: Cannot invoke "choreo.trajectory.Trajectory.sampleAt(double, boolean)" because "this.m_trajectory" is null
	at com.team973.frc2025.subsystems.composables.DriveWithTrajectory.getOutput(DriveWithTrajectory.java:117)
	at com.team973.frc2025.subsystems.DriveController.update(DriveController.java:158)
	at com.team973.frc2025.subsystems.swerve.GreyPoseEstimator.observeOdometryData(GreyPoseEstimator.java:86)
	at com.team973.frc2025.subsystems.swerve.OdometrySupplier.doCycle(OdometrySupplier.java:173)
	at com.team973.frc2025.subsystems.swerve.OdometrySupplier.doRun(OdometrySupplier.java:150)
	at com.team973.frc2025.subsystems.swerve.OdometrySupplier.run(OdometrySupplier.java:128)
	at java.base/java.lang.Thread.run(Unknown Source)
Error at com.team973.frc2025.subsystems.composables.DriveWithTrajectory.getOutput(DriveWithTrajectory.java:117): Cannot invoke "choreo.trajectory.Trajectory.sampleAt(double, boolean)" because "this.m_trajectory" is null
	at com.team973.frc2025.subsystems.composables.DriveWithTrajectory.getOutput(DriveWithTrajectory.java:117)
	at com.team973.frc2025.subsystems.DriveController.update(DriveController.java:158)
	at com.team973.frc2025.subsystems.swerve.GreyPoseEstimator.observeOdometryData(GreyPoseEstimator.java:86)
	at com.team973.frc2025.subsystems.swerve.OdometrySupplier.doCycle(OdometrySupplier.java:173)
	at com.team973.frc2025.subsystems.swerve.OdometrySupplier.doRun(OdometrySupplier.java:150)
	at com.team973.frc2025.subsystems.swerve.OdometrySupplier.run(OdometrySupplier.java:128)
	at java.base/java.lang.Thread.run(Unknown Source)
```

The problem was that we were setting the DriveWithTrajectory controller
to active before we initialized its state. With this change we swap the
order so we initialize state before we set it to active.