armivator: Do not zero on startup
I think the elevator was pushing down with considerable force because
it was being reset to 0.25" height at bootup when in reality it was
at 0.0" height.

With this change, we do not zero on bootup --- we only zero when we
are sure that the hall sensor has transitioned from false to true.